removed apport (damn thing swalling the farking core file!!!) ... will resume this l8r.
(gdb) bt
    rad_exp=rad_exp@entry=0x7f127cdd3a90, x=x@entry=0x7f127cdd43b8,
n=-104, more=0) at get_str.c:335

+++ src/arblib/functions/SineFunction.java
@@ -25,7 +25,7 @@ public class SineFunction implements
       assert order == 1;
       // System.out.format( "THe z addy is 0x%x", z.swigCPtr );
       Real sin = z.sin(prec, w);
-      System.out.format( "z=%s w=%s\n", z.doubleValue(), w.doubleValue());
+      System.out.format( "z=%s w=%s\n", z, w);
 //      System.out.flush();
       return sin;
     }
